src/requirements.txt
amqp==5.0.6
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.6.0
billiard==3.6.4.0
boto==2.49.0
boto3==1.18.2
botocore==1.21.2
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.2
certifi==2024.7.4
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.3
colorama==0.4.4
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
distlib==0.3.2
distro==1.5.0
Django==4.2.15
django-admin-list-filter-dropdown==1.0.3
django-axes[ipware]==6.4.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-grappelli==3.0.8
django-json-widget==1.1.1
django-rest-swagger==2.2.0
django-seo-js==0.4.1
django-ses==3.5.0
django-storages==1.11.1
django-webpack-loader==1.1.0
djangorestframework==3.15.2
djangorestframework-gis==0.18
docutils==0.17.1
enum-compat==0.0.3
future==0.18.3
google-auth==1.33.0
google-auth-oauthlib==0.4.4
gspread==3.7.0
html5lib==1.1
httplib2==0.19.1
idna==3.7
importlib-metadata==4.6.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==3.1.4
jmespath==0.10.0
kombu==5.1.0
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==2.0.1
msgpack==1.0.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.0
pep517==0.11.0
Pillow==10.3.0
progress==1.5
promise==2.3
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.0
resolvelib==0.5.4
rsa==4.7.2
Rx==3.2.0
s3transfer==0.5.0
simplejson==3.17.3
singledispatch==3.6.2
six==1.16.0
sqlparse==0.5.0
toml==0.10.2
tomli==1.0.4
typing-extensions==3.10.0.0
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.19
vine==5.0.0
watchtower==3.0.0
webencodings==0.5.1
zipp==3.19.1
